requirements.txt
streamlit==1.42.0
supabase==2.13.0
requests==2.32.3
pydantic==2.10.6
python-dateutil==2.9.0.post0
